requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# AI and ML
openai==1.3.7
tiktoken==0.5.2
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
python-magic==0.4.27

# File processing
aiofiles==23.2.1
boto3==1.34.0
google-cloud-storage==2.10.0
dropbox==11.36.2
requests==2.31.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
